Refactor existing classes
Move most of the short method definitions into header files.
This made some of the cpp's empty, but I'd prefer to keep
them in a source tree as a reminder that they are real classes,
not interfaces.
Improve PolandBall::Entity class, make it more fail-safe
with default collider and sprite ready to use.

Note: This commit is too messy and should be splitted into two
ones. I must stop coding under the meth...